requirements.txt.txt
# ── Web framework ─────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.27.1

# ── Async HTTP requests ───────────────────────
httpx==0.27.0

# ── Twilio Messaging ──────────────────────────
twilio==9.0.4

# ── Postgres DB Drivers ───────────────────────
asyncpg==0.29.0      # for FastAPI
psycopg[binary]==3.1.18  # for Streamlit dashboard

# ── Geocoding + Math ──────────────────────────
geopy==2.4.1         # optional, if haversine wasn't hand-written
# math module is built-in, no need to include

# ── Streamlit Dashboard ───────────────────────
streamlit==1.35.0
streamlit-folium==0.18.0
folium==0.16.0
pandas==2.2.2

# ── Env loading ───────────────────────────────
python-dotenv==1.0.1

# ── Dev only ──────────────────────────────────
# pytest==8.2.1     # optional if you add test cases
